requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.16.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.7
django-js-asset==3.1.2
django-mptt==0.16.0
docxcompose==1.4.0
docxtpl==0.19.0
filelock==3.15.4
fonttools==4.57.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
oscrypto==1.3.0
packaging==24.2
pillow==11.2.1
platformdirs==4.2.2
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-docx==1.1.2
PyYAML==6.0.2
qrcode==8.1
reportlab==4.4.0
requests==2.32.3
setuptools==75.8.2
six==1.16.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
virtualenv==20.26.3
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
